requirements.txt
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
docopt==0.6.2
docutils==0.20.1
Faker==17.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jaraco.classes==3.2.3
jsonpath==0.82
keyring==23.13.1
Markdown==3.4.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
pipreqs==0.4.13
pkginfo==1.9.6
pyasn1==0.4.8
Pygments==2.15.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.30.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
rsa==4.7.2
six==1.16.0
twine==4.0.2
typing-extensions==4.5.0
urllib3==2.0.2
webencodings==0.5.1
yarg==0.1.9
zipp==3.15.0
DingtalkChatbot==1.5.7
Jinja2==3.1.2
PyYAML==6.0